pyproject.toml
[tool.poetry]
name = "css"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7, <3.9"
pyppeteer = "^1.0.2"
"ruamel.yaml" = "^0.17.20"
discord = "^1.7.3"
click = "^8.0.3"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"